### 2023-03-07 #5229 @editorialbot
[PRE REVIEW]: QuaC: A Pipeline Implementing Quality Control Best Practices for Genome Sequencing and Exome Sequencing Data
**Submitting author:** @ManavalanG (<a href="http://orcid.org/0000-0002-8606-0113">Manavalan Gajapathy</a>)
**Repository:** https://github.com/uab-cgds-worthey/quac
**Branch with paper.md** (empty if default branch): joss_manuscript
**Version:** 1.3
