### 2023-03-29 #5229 @Redmar-van-den-Berg
I'm trying to follow along with https://quac.readthedocs.io/en/stable/system_testing/, and I've run into 2 issues
1. The pipeline does not create the output and log folders automatically, so the user has to create the folders manually before the pipeline is started
2. There is no automatic way to fetch the required reference files that are described in configs/workflow.yaml
