### 2023-03-31 uab-cgds-worthey/quac#65 @ManavalanG
Create a script to retrieve necessary datasets for workflow configs
As requested [here](https://github.com/openjournals/joss-reviews/issues/5313#issuecomment-1490213554)
