### 2023-04-02 a05c30e
adds script to set up dependency datasets - #65
