### 2023-04-02 086e06b
adds workflow config for system testing
