### 2023-04-02 b431241
updates testing doc about dataset dependencies
