### 2023-04-02 839397c
updates docs on dataset dependencies - closes #65
