### 2023-04-02 #5313 @ManavalanG
@Redmar-van-den-Berg Thanks for accepting to review our pipeline and providing us an initial feedback.

>The pipeline does not create the output and log folders automatically, so the user has to create the folders manually before the pipeline is started

We chose this setup deliberately as this acts as a sanity check for us to be certain that the output and log directories were chosen correctly by the user.

>There is no automatic way to fetch the required reference files that are described in configs/workflow.yaml

Good point. I have now automated retrieval and processing of the necessary datasets. [Docs are updated](https://github.com/uab-cgds-worthey/quac/issues/65) to reflect this change, and workflow-config is now available specifically for system testing (`.test/configs/test_workflow.yaml`).

PS- Please be sure to use `latest` version of the docs. This can be chosen by clicking version button on the bottom right in the docs website.
