### 2023-04-03 cec28ca
moves script to src
