### 2023-04-03 b000c42
updates workflow config to use retrieved datasets
