+++ .test/configs/test_workflow.yaml
@@ -1,32 +1,0 @@
-datasets:
-  # path to ref genome
-  ref: ".test/dependency_datasets/reference_genome/GCA_000001405.15_GRCh38_no_alt_analysis_set.fna"
-  # path to Somalier dataset dependencies. Available at https://github.com/brentp/somalier
-  somalier:
-    sites: ".test/dependency_datasets/somalier/sites.hg38.vcf.gz"
-    labels_1kg: ".test/dependency_datasets/somalier/ancestry-labels-1kg.tsv"
-    somalier_1kg: ".test/dependency_datasets/somalier/1kg-somalier/"
-  # path to verifyBamID dataset dependencies
-  verifyBamID:
-    # Refer to https://github.com/Griffan/VerifyBamID for downloading this file
-    svd_dat_wgs: ".test/dependency_datasets/verifyBamID/1000g.phase3.100k.b38.vcf.gz.dat"
-    # Refer to QuaC docs to create this file. See section "Prepare verifybamid datasets for exome mode" in how to run quac
-    svd_dat_exome: ".test/dependency_datasets/verifyBamID/exome/1000g.phase3.10k.b38_chr.exome.vcf.gz.dat"
-
-#### hardware resources ####
-resources:
-  qualimap_bamqc:
-    no_cpu: 1
-    mem_per_cpu: "4G"
-  mosdepth_coverage:
-    no_cpu: 1
-  verifybamid:
-    no_cpu: 1
-
-#### slurm partitions ####
-slurm_partitions:
-  # modify below to provide partitions available in your slurm and their time limits
-  express: "02:00:00"
-  short: "12:00:00"
-  medium: "50:00:00"
-  long: "150:00:00"
