### 2023-04-03 0fbf637
updates docs to reflect workflow config changes
