### 2023-04-03 977f474
fixes outdirpath
