### 2023-04-03 #5313 @Redmar-van-den-Berg
> We chose this setup deliberately as this acts as a sanity check for us to be certain that the output and log directories were chosen correctly by the user.

I don't think this adds a lot of safety, and it does make it more difficult to run the pipeline.

> Good point. I have now automated retrieval and processing of the necessary datasets. [Docs are updated](https://quac.readthedocs.io/en/latest/reqts_configs/#dependency-datasets) to reflect this change, and workflow-config is now updated to utilize the datasets downloaded using above script.

The `.test/setup_test_datasets.sh` does not run on my system. It uses the `module` system, which is not available in the `quac` conda environment. It also refers to `/data/project/worthey_lab/samples/NA12878`, which will not exist on most  users systems. Would it be possible to include a small example bam file in the repository for testing purposes? That way, QuaC could also use automated tests on github, which is one of the recommendations from JOSS.
