### 2023-04-03 #5313 @ManavalanG
>The .test/setup_test_datasets.sh does not run on my system.

We make output of this script readily available in the repo. Please see `.test/ngs-data/test_project/analysis` for the output. This includes bam, vcf, etc. `.test/setup_test_datasets.sh` was included to show the steps taken to create the files used in system testing, and we do not expect most users to execute the script :)
