### 2023-04-04 d925099
creates dirpath if user provided dirs dont exist - #66
