### 2023-04-04 e9991d9
retires cheaha specific env variable - #66
