+++ src/run_quac.py
@@ -30,7 +30,7 @@ def make_dir(d):
 
 def read_workflow_config(workflow_config_fpath):
     """
-    Read workflow config file to 
+    Read workflow config file to
     (1) identify paths to be mounted for singularity.
     (2) get slurm partitions.
     """
@@ -50,7 +50,7 @@ def read_workflow_config(workflow_config_fpath):
     # verifyBamID resource files
     for resource in datasets["verifyBamID"]:
         mount_paths.add(Path(datasets["verifyBamID"][resource]).parent)
-        
+
     # get slurm partitions
     slurm_partitions_dict = data["slurm_partitions"]
 
@@ -58,7 +58,13 @@ def read_workflow_config(workflow_config_fpath):
 
 
 def gather_mount_paths(
-    projects_path, project_name, pedigree_path, out_dir, log_dir, quac_watch_config, workflow_config
+    projects_path,
+    project_name,
+    pedigree_path,
+    out_dir,
+    log_dir,
+    quac_watch_config,
+    workflow_config,
 ):
     """
     Returns paths that need to be mounted to singularity
@@ -140,7 +146,7 @@ def create_snakemake_command(args, repo_path, mount_paths):
             "--cluster 'sbatch --ntasks {cluster.ntasks} --partition {cluster.partition}"
             " --cpus-per-task {cluster.cpus-per-task} --mem-per-cpu {cluster.mem-per-cpu}"
             " --job-name {cluster.jobname} --output {cluster.output} --parsable'",
-        ]        
+        ]
 
     # add any user provided extra args for snakemake
     if args.extra_args:
@@ -185,7 +191,7 @@ def main(args):
     _, slurm_partition_times = read_workflow_config(args.workflow_config)
 
     slurm_resources = {
-        "partition": args.slurm_partition, 
+        "partition": args.slurm_partition,
         "ntasks": "1",
         "time": slurm_partition_times[args.slurm_partition],
         "cpus-per-task": "1" if args.subtasks_slurm else "4",
@@ -264,8 +270,10 @@ if __name__ == "__main__":
     )
     WORKFLOW.add_argument(
         "--quac_watch_config",
-        help=("YAML config path specifying QC thresholds for QuaC-Watch." 
-              " See directory 'configs/quac_watch/' in quac repo for the included config files."),
+        help=(
+            "YAML config path specifying QC thresholds for QuaC-Watch."
+            " See directory 'configs/quac_watch/' in quac repo for the included config files."
+        ),
         type=lambda x: is_valid_file(PARSER, x),
         metavar="",
     )
@@ -311,7 +319,7 @@ if __name__ == "__main__":
         "--subtasks_slurm",
         action="store_true",
         help="Flag indicating that the main Snakemake process of QuaC should submit subtasks of"
-        " the workflow as Slurm jobs instead of running them on the same machine as itself"
+        " the workflow as Slurm jobs instead of running them on the same machine as itself",
     )
 
     ############ Args for QuaC wrapper tool  ############
@@ -351,9 +359,9 @@ if __name__ == "__main__":
     WRAPPER.add_argument(
         "--snakemake_slurm",
         action="store_true",
-        help="Flag indicating that the main Snakemake process of QuaC should be" 
-        " submitted to run in a Slurm job instead of executing in the current" 
-        " environment. Useful for headless execution on Slurm-based HPC systems."
+        help="Flag indicating that the main Snakemake process of QuaC should be"
+        " submitted to run in a Slurm job instead of executing in the current"
+        " environment. Useful for headless execution on Slurm-based HPC systems.",
     )
     RERUN_FAILED_DEFAULT = 1
     WRAPPER.add_argument(
[…]
