### 2023-04-04 f36968b
updates doc on refactoring of $USER_SCRATCH
