+++ configs/workflow.yaml
@@ -3,17 +3,17 @@
 
 datasets:
   # path to ref genome
-  ref: ".test/dependency_datasets/reference_genome/GCA_000001405.15_GRCh38_no_alt_analysis_set.fna"
+  ref: "data/external/dependency_datasets/reference_genome/GCA_000001405.15_GRCh38_no_alt_analysis_set.fna"
   # path to Somalier dataset dependencies. Available at https://github.com/brentp/somalier
   somalier:
-    sites: ".test/dependency_datasets/somalier/sites.hg38.vcf.gz"
-    labels_1kg: ".test/dependency_datasets/somalier/ancestry-labels-1kg.tsv"
-    somalier_1kg: ".test/dependency_datasets/somalier/1kg-somalier/"
+    sites: "data/external/dependency_datasets/somalier/sites.hg38.vcf.gz"
+    labels_1kg: "data/external/dependency_datasets/somalier/ancestry-labels-1kg.tsv"
+    somalier_1kg: "data/external/dependency_datasets/somalier/1kg-somalier/"
   # path to verifyBamID dataset dependencies. Available at https://github.com/Griffan/VerifyBamID
   verifyBamID:
-    svd_dat_wgs: ".test/dependency_datasets/verifyBamID/1000g.phase3.100k.b38.vcf.gz.dat"
+    svd_dat_wgs: "data/external/dependency_datasets/verifyBamID/1000g.phase3.100k.b38.vcf.gz.dat"
     # retrieved from https://github.com/Griffan/VerifyBamID and 'chr' prefix added to the contigs
-    svd_dat_exome: ".test/dependency_datasets/verifyBamID/exome/1000g.phase3.10k.b38_chr.exome.vcf.gz.dat"
+    svd_dat_exome: "data/external/dependency_datasets/verifyBamID/exome/1000g.phase3.10k.b38_chr.exome.vcf.gz.dat"
 
 #### hardware resources ####
 resources:
