### 2023-04-04 #5313 @ManavalanG
>Ideally, QuaC would include either a reduced test version of these files, or a script to fetch them automatically.

QuaC includes script [src/setup_dependency_datasets.sh](https://github.com/uab-cgds-worthey/quac/blob/joss_manuscript/src/setup_dependency_datasets.sh) to fetch all the datasets listed under `datasets` in [configs/workflow.yaml](https://github.com/uab-cgds-worthey/quac/blob/joss_manuscript/configs/workflow.yaml). If those are not the files you were referring to, please let me know.

I cleaned up [the docs](https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file) a bit for clarity on this topic. I hope this helps.
