### 2023-04-05 4a5b76a
removes default for --projects_path
