### 2023-04-05 #5313 @Redmar-van-den-Berg
Thanks for pointing me to the script, I was able to download the required reference files!

However, I run into a crash when I try to run the pipeline itself, following along to the [documentation](https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file).

```bash
 python src/run_quac.py       --project_name test_project       --projects_path ".test/ngs-data/"       --pedigree ".test/configs/${PRIOR_QC_STATUS}/${PROJECT_CONFIG}.ped"       --outdir "$USER_SCRATCH/tmp/quac/results/test_${PROJECT_CONFIG}_exome-${PRIOR_QC_STATUS}/analysis"       --quac_watch_config "configs/quac_watch/exome_quac_watch_config.yaml"       --exome       $USE_SLURM       -e="--conda-create-envs-only"
########################################
Command to run the pipeline:
snakemake \
	--snakefile '/home/username/devel/quac/workflow/Snakefile' \
	--config project_name='test_project' projects_path='/home/username/devel/quac/.test/ngs-data' ped='/home/username/devel/quac/.test/configs/no_priorQC/project_2samples.ped' quac_watch_config='/home/username/devel/quac/configs/quac_watch/exome_quac_watch_config.yaml' workflow_config='/home/username/devel/quac/configs/workflow.yaml' unique_id='2ad58efa-8557-4807-b85d-ed1d4752ff2f' out_dir='/home/username/devel/quac/scratch/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis' log_dir='/home/username/devel/quac/scratch/tmp/quac/logs' exome='True' include_prior_qc_data='False' allow_sample_renaming='False' \
	--restart-times 1 \
	--use-conda \
	--use-singularity \
	--singularity-args '--cleanenv --bind /home/username/devel/quac/scratch/tmp/quac/tmp:/tmp --bind /home/username/devel/quac/.test/ngs-data/test_project/analysis,/home/username/devel/quac/scratch/tmp/quac/logs,/home/username/devel/quac/scratch/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis,data/external/dependency_datasets/somalier,data/external/dependency_datasets/verifyBamID,data/external/dependency_datasets/reference_genome,/home/username/devel/quac/.test/configs/no_priorQC,/home/username/devel/quac/configs/quac_watch/exome_quac_watch_config.yaml,data/external/dependency_datasets/verifyBamID/exome' \
	--profile '/home/username/devel/quac/src/slurm/slurm_profile' \
	--conda-create-envs-only
########################################

Slurm job name   : "quac-2023-04-05T10.10.34.732874"
Slurm job script : "/home/username/devel/quac/scratch/tmp/quac/logs/quac-2023-04-05T10.10.34.732874.sh"
// Processing project: test_project
// Project path: "/home/username/devel/quac/.test/ngs-data/test_project/analysis"
// Exome mode: True
// Include prior QC data: False
// WARNING: '.test' present in the path supplied via --projects_path. So testing mode is used.
Building DAG of jobs...
Pulling singularity image docker://continuumio/miniconda3:4.7.12.
Pulling singularity image docker://brentp/somalier:v0.2.13.
Creating conda environment configs/env/quac_watch.yaml...
Downloading and installing remote packages.
CreateCondaEnvironmentException:
Could not create conda environment from /home/username/devel/quac/configs/env/quac_watch.yaml:
FATAL:   container creation failed: unable to add /home/username/devel/quac/data/external/dependency_datasets/somalier to mount list: destination must be an absolute path

  File "/home/username/miniconda3/envs/quac/lib/python3.6/site-packages/snakemake/deployment/conda.py", line 389, in create
Traceback (most recent call last):
  File "src/run_quac.py", line 372, in <module>
    main(ARGS)
  File "src/run_quac.py", line 202, in main
    submit_slurm_job(pipeline_cmd, job_dict)
  File "/home/username/devel/quac/src/slurm/submit_slurm_job.py", line 60, in submit_slurm_job
    job_id = slurm_job.run(**params_dict)
  File "/home/username/miniconda3/envs/quac/lib/python3.6/site-packages/slurmpy/slurmpy.py", line 171, in run
    res = subprocess.check_output(args).strip()
  File "/home/username/miniconda3/envs/quac/lib/python3.6/subprocess.py", line 356, in check_output
    **kwargs).stdout
  File "/home/username/miniconda3/envs/quac/lib/python3.6/subprocess.py", line 438, in run
    output=stdout, stderr=stderr)
subprocess.CalledProcessError: Command '['bash', '/home/username/devel/quac/scratch/tmp/quac/logs/quac-2023-04-05T10.10.34.732874.sh']' returned non-zero exit status 1.
```
