### 2023-04-05 #5313 @ManavalanG
@Redmar-van-den-Berg

>The pipeline does not create the output and log folders automatically, so the user has to create the folders manually before the pipeline is started

We have now refactored QuaC to auto-create these folders (https://github.com/uab-cgds-worthey/quac/pull/67). Please see here for updated docs - https://quac.readthedocs.io/en/latest/
