### 2023-04-05 uab-cgds-worthey/quac#68 @wilkb777
@Redmar-van-den-Berg this is a tricky issue, it's a case of Snakemake wanting to create a Singularity image using Anaconda inside the image to build an environment and it's trying to mount the somalier data directory as part of the process.

Can you tell me if any part of the path `/home/username/devel/quac` is a symlink or some sort of non-absolute path on your setup? Since this is an issue of trying to mount (or bind) a host path to the inside of the container and the error says the path looks like an absolute path but singularity doesn't think so this issue sounds a lot like what a user was seeing in this [SO post](https://stackoverflow.com/a/56931352).

Perhaps you could run

```shell
realpath ~
```

and tell us if the path it returns is the same as `/home/username`? On the HPC system I have access to the home directory is not truly `/home/username` it ends up looking like this

```shell
username@login004:~$ pwd
/home/username

username@login004:~$ realpath ~
/data/user/home/username
```

Knowing that info would be helpful in determining if moving the repo to a non-home directory location would help.

In the past I've seen issues with Singularity and mounting paths in a user's home directory on top of the typical default mount that Singularity does where it auto-mounts a users home directory automatically. It *shouldn't* be an issue but for reasons I never figured out in the past it ended up being resolved by working out of a non-home directory location. I'm not sure that this is the case here but might be something to try.
