### 2023-04-06 uab-cgds-worthey/quac#68 @Redmar-van-den-Berg
As far as I can tell, my home directory is a regular path:
```bash
$ realpath ~
/home/username
```

Rather than trying to debug this, would it be possible for QuaC to switch over to using singularity images instead of conda? Conda environments tend to become uninstallable over time, which means pipelines using them are not fully reproducible.

I tend to use the Biocontainers hosted on quay.io, which are built automatically for all conda packages (e.g. https://quay.io/repository/biocontainers/goleft, which is available as `docker://quay.io/biocontainers/goleft:0.2.4--0`). It is also possible to automatically build multi package containers if you need multiple tools in a single workflow step.
