### 2023-04-10 4844ae8
removes unused conda envs
