+++ configs/env/quac_watch.yaml
@@ -9,80 +9,80 @@ dependencies:
   - jinja2==2.11.3
   - pyyaml=5.3
   - fire=0.3
-  - multiqc=1.9
-  # other dependencies
-  - _libgcc_mutex=0.1
-  - _openmp_mutex=4.5
-  - brotlipy=0.7.0
-  - ca-certificates=2021.5.30
-  - certifi=2021.5.30
-  - cffi=1.14.6
-  - chardet=4.0.0
-  - charset-normalizer=2.0.0
-  - click=8.0.1
-  - coloredlogs=15.0.1
-  - colormath=3.0.0
-  - cryptography=3.4.7
-  - cycler=0.10.0
-  - decorator=5.0.9
-  - freetype=2.10.4
-  - future=0.18.2
-  - humanfriendly=9.2
-  - idna=3.1
-  - importlib-metadata=4.6.3
-  - jbig=2.1
-  - jpeg=9d
-  - kiwisolver=1.3.1
-  - lcms2=2.12
-  - ld_impl_linux-64=2.36.1
-  - lerc=2.2.1
-  - libblas=3.9.0
-  - libcblas=3.9.0
-  - libdeflate=1.7
-  - libffi=3.3
-  - libgcc-ng=11.1.0
-  - libgfortran-ng=11.1.0
-  - libgfortran5=11.1.0
-  - libgomp=11.1.0
-  - liblapack=3.9.0
-  - libopenblas=0.3.17
-  - libpng=1.6.37
-  - libstdcxx-ng=11.1.0
-  - libtiff=4.3.0
-  - libwebp-base=1.2.0
-  - lz4-c=1.9.3
-  - lzstring=1.0.4
-  - markdown=3.3.4
-  - markupsafe=2.0.1
-  - matplotlib-base=3.3.4
-  - ncurses=6.2
-  - networkx=2.5
-  - olefile=0.46
-  - openjpeg=2.4.0
-  - openssl=1.1.1k
-  - pillow=8.3.1
-  - pip=21.2.3
-  - pycparser=2.20
-  - pyopenssl=20.0.1
-  - pyparsing=2.4.7
-  - pysocks=1.7.1
-  - python-dateutil=2.8.2
-  - python_abi=3.6
-  - pytz=2021.1
-  - readline=8.1
-  - requests=2.26.0
-  - setuptools=49.6.0
-  - simplejson=3.17.3
-  - six=1.16.0
-  - spectra=0.0.11
-  - sqlite=3.36.0
-  - termcolor=1.1.0
-  - tk=8.6.10
-  - tornado=6.1
-  - typing_extensions=3.10.0.0
-  - urllib3=1.26.6
-  - wheel=0.37.0
-  - xz=5.2.5
-  - yaml=0.2.5
-  - zipp=3.5.0
-  - zlib=1.2.11
+  # - multiqc=1.9
+  # # other dependencies
+  # - _libgcc_mutex=0.1
+  # - _openmp_mutex=4.5
+  # - brotlipy=0.7.0
+  # - ca-certificates=2021.5.30
+  # - certifi=2021.5.30
+  # - cffi=1.14.6
+  # - chardet=4.0.0
+  # - charset-normalizer=2.0.0
+  # - click=8.0.1
+  # - coloredlogs=15.0.1
+  # - colormath=3.0.0
+  # - cryptography=3.4.7
+  # - cycler=0.10.0
+  # - decorator=5.0.9
+  # - freetype=2.10.4
+  # - future=0.18.2
+  # - humanfriendly=9.2
+  # - idna=3.1
+  # - importlib-metadata=4.6.3
+  # - jbig=2.1
+  # - jpeg=9d
[…]
