### 2023-04-10 69ba630
removes quac-watch conda env
