### 2023-04-10 c1d5bec
updates docs on env type created
