### 2023-04-10 uab-cgds-worthey/quac#68 @wilkb777
That's a good idea, I know this is the preference of setup for Nextflow NF-Core based pipelines and @ManavalanG had dealt with a scenario in #52 that had prompted the thought to move move away completely from conda + singularity to just singularity. I'm reviewing #69 right now as part of this effort.

I'm still really perplexed by this issue where Singularity can't mount a directory into the container. @Redmar-van-den-Berg if you have time before #69 is reviewed and merged (and this is totally at your discretion b/c it's more debugging and more potentially unnecessary work on you) could you sanity check the mounting setup for Singularity?

If you have time could you quickly check that the `lolcow` testing container can bind those mount paths and print out the cow to the terminal?

The command would be

```shell
cd /home/username/devel/quac
singularity run --cleanenv --bind /home/username/devel/quac/scratch/tmp/quac/tmp:/tmp --bind /home/username/devel/quac/.test/ngs-data/test_project/analysis,/home/username/devel/quac/scratch/tmp/quac/logs,/home/username/devel/quac/scratch/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis,data/external/dependency_datasets/somalier,data/external/dependency_datasets/verifyBamID,data/external/dependency_datasets/reference_genome,/home/username/devel/quac/.test/configs/no_priorQC,/home/username/devel/quac/configs/quac_watch/exome_quac_watch_config.yaml,data/external/dependency_datasets/verifyBamID/exome docker://sylabsio/lolcow
```

The only thing I could think is something is preventing the singularity command from correctly binding relative paths, in particular `data/external/dependency_datasets/somalier` since it's the one giving the error and the first relative bind path listed in your log command. If that works fine and prints out the cow like

```
 ______________________________
< Mon Apr 10 15:14:40 CDT 2023 >
 ------------------------------
        \   ^__^
         \  (oo)\_______
            (__)\       )\/\
                ||----w |
                ||     ||
```
then I have no idea the issue and hope that just moving strictly to containers resolves whatever issue is coming up with the combo of conda + singularity 😂 . I completely understand if your preference is to just wait for the PR switching to full containers is merged and test that setup directly.
