### 2023-04-14 fac42b8
removes conda-create-envs-only section in docs
