+++ docs/reqts_configs.md
@@ -46,40 +46,3 @@ QuaC requires a workflow config file in yaml format (default: `configs/workflow.
     Custom workflow config file can be provided to QuaC via `--workflow_config`.
 
 
-## Create singularity environments for tools used in QuaC pipeline
-
-All the jobs initiated by QuaC's snakemake workflow would be run in Singularity environment. It may be a good idea to create these environments before they are run with actual samples. While this step is
-optional, this will ensure that there will not be any conflicts when running multiple instances of the pipeline.
-
-Running the commands below will create the Singularity environments. Note that it will exit right
-after creating the environments, and it will not run any QC analyses on the input samples provided.
-
-```sh
-# For Cheaha users only. Set up environment. 
-module reset
-module load Anaconda3/2020.02
-module load Singularity/3.5.2-GCC-5.4.0-2.26
-
-# activate conda env
-conda activate quac
-
-PROJECT_CONFIG="project_2samples"
-PRIOR_QC_STATUS="no_priorQC"
-
-USE_SLURM="--snakemake_slurm --subtasks_slurm"
-# USE_SLURM=""  # uncomment and use this out if you don't want to use slurm at all
-
-python src/run_quac.py \
-      --project_name test_project \
-      --projects_path ".test/ngs-data/" \
-      --pedigree ".test/configs/${PRIOR_QC_STATUS}/${PROJECT_CONFIG}.ped" \
-      --outdir "data/quac/results/test_${PROJECT_CONFIG}_exome-${PRIOR_QC_STATUS}/analysis" \
-      --quac_watch_config "configs/quac_watch/exome_quac_watch_config.yaml" \
-      --exome \
-      $USE_SLURM \
-      -e="--conda-create-envs-only"
-```
-
-!!! info
-
-    In Cheaha, it took 25 mins to complete this task. Resources used for the task: 1 CPU and 8GB memory.
