### 2023-04-14 012884b
adds optional step to pull singularity images
