### 2023-04-24 #5313 @ManavalanG
@Redmar-van-den-Berg @brentp I would like to inform that we have refactored QuaC to move away from creation of conda environment within singularity containers. QuaC now simply depends on webserver-hosted container images, retrieves them automatically as part of the pipeline and uses them for snakemake-initiated jobs (#70).
