### 2023-04-26 f77a06a
replaces with output dir structure for include_priorQC mode
