+++ docs/system_testing.md
@@ -101,51 +101,54 @@ python src/run_quac.py \
 Output directory structure for WGS + `include_prior_qc` mode would look like this.
 
 ```sh
-$ tree -d data/quac/results/test_project_2samples_wgs-no_priorQC/ -L 5
-data/quac/results/test_project_2samples_wgs-no_priorQC/
+$ tree data/quac/results/test_project_2samples_wgs-include_priorQC/ -L 5
+data/quac/results/test_project_2samples_wgs-include_priorQC/
 └── analysis
-    ├── C
+    ├── A
     │   └── qc
     │       ├── bcftools-index
-    │       │   └── ...
+    │       │   └── ...
     │       ├── bcftools-stats
-    │       │   └── ...
+    │       │   └── ...
     │       ├── mosdepth
-    │       │   └── ...
+    │       │   └── ...
     │       ├── multiqc_final_pass
-    │       │   ├── ...
-    │       │   └── C_multiqc.html
+    │       │   ├── ...
+    │       │   └── A_multiqc.html
     │       ├── multiqc_initial_pass
-    │       │   ├── ...
-    │       │   └── C_multiqc.html
+    │       │   ├── ...
+    │       │   └── A_multiqc.html
     │       ├── picard-stats
-    │       │   └── ...
+    │       │   └── ...
     │       ├── quac_watch
-    │       │   └── ...
+    │       │   └── ...
     │       ├── qualimap
-    │       │   └── ...
+    │       │   └── ...
     │       ├── samtools-stats
-    │       │   └── ...
+    │       │   └── ...
     │       └── verifyBamID
-    │           └── ...
-    ├── D
+    │           └── ...
+    ├── B
     │   └── qc
-    │       └── same directory structure as that of sample C
+    │       └── same directory structure as that of sample A
     └── project_level_qc
         ├── covviz
-        │   └── ...
+        │   └── ...
         ├── indexcov
-        │   └── ...
+        │   └── ...
         ├── mosdepth
-        │   └── ...
+        │   └── ...
         ├── multiqc
-        │   ├── ...
+        │   ├── configs
+        │   │   └── aggregated_rename_configs.tsv
+        │   ├── multiqc_report_data
+        │   │   └── ...
         │   └── multiqc_report.html
         └── somalier
             ├── ancestry
-            │   └── ...
+            │   └── ...
             ├── extract
-            │   └── ...
+            │   └── ...
             └── relatedness
                 └── ...
 ```
