+++ docs/installation.md
@@ -5,10 +5,10 @@ Installation involves
 - fetching the source code
 - creating the conda environment
 
-Quac is expected to work in linux environments.
-
 ## Requirements
 
+- Linux OS
+  - Tested in Red Hat Enterprise Linux  v7.9
 - Git
 - Anaconda/miniconda
     - Tested with Anaconda3/2020.02
