### 2023-05-09 uab-cgds-worthey/quac#68 @Redmar-van-den-Berg
The issue still happens for me, even if I run the latest version of the pipeline that does not use Conda.

I think it is related to the snakemake command attempting to bind mount a whole bunch of folders that do not exist:
```bash
--singularity-args '--cleanenv --bind /data/lumc/devel/quac/data/quac/tmp:/tmp --bind /data/lumc/devel/quac/configs/quac_watch/wgs_quac_watch_config.yaml,data/external/dependency_datasets/somalier,/data/lumc/devel/quac/.test/configs/no_priorQC,data/external/dependency_datasets/reference_genome,data/external/dependency_datasets/verifyBamID/exome,/data/lumc/devel/quac/data/quac/results/test_project_2samples_wgs-no_priorQC/analysis,/data/lumc/devel/quac/data/quac/logs,/data/lumc/devel/quac/.test/ngs-data/test_project/analysis,data/external/dependency_datasets/verifyBamID
```
