### 2023-05-09 4a4bab7
verifies mount paths exist
