### 2023-05-09 3fe8a3f
separates path check component
