### 2023-05-09 2ae1e1e
verifies actual filepath instead of its parent dirpath
