### 2023-05-11 e81549a
refactors to verify all mount paths at once
