### 2023-05-11 bcb240c
uses resolve instead of abs as it can deal with symlinks
