### 2023-05-11 7273e4b
resolves datasets path in workflow config
