### 2023-05-12 #5313 @brentp
@ManavalanG would you show how to install the software and run on a fresh system?

If possible, I also suggest that you have someone else install and run the example without any guidance from you, perhaps while you watch. Please do this outside of your usual cluster.  You have spent much time documenting and building this, but I also am not able to get it running. Currently, I see:

```
$ python src/run_quac.py       --project_name test_project       --projects_path ".test/ngs-data/"       --pedigree ".test/configs/${PRIOR_QC_STATUS}/${PROJECT_CONFIG}.ped"       --outdir "$USER_SCRATCH/tmp/quac/results/test_${PROJECT_CONFIG}_exome-${PRIOR_QC_STATUS}/analysis"       --quac_watch_config "configs/quac_watch/exome_quac_watch_config.yaml"       --exome       $USE_SLURM
Created directory: /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis
ERROR: Following directories that are part of your input were not found:
/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/somalier
/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/reference_genome
/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/verifyBamID
/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/verifyBamID/exome
```
It's not clear to me how to resolve this.
