### 2023-05-12 #5313 @ManavalanG
@brentp It appears that datasets specified in the workflow config file `configs/quac_watch/exome_quac_watch_config.yaml` are not available in your system. Please see [directions here](https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file) on using the included script to set up all necessary dataset dependencies. Please let us know if this doesn't resolve the issue.
