### 2023-05-12 0c0dc61
improves error message when mount paths not exist
