### 2023-05-12 #5313 @ManavalanG
@brentp

>If possible, I also suggest that you have someone else install and run the example without any guidance from you, perhaps while you watch. Please do this outside of your usual cluster.

Thanks for the suggestion! We have had a user test the installation, but they used the same hardware resources as we use.

PS - I modified the error message now and then highlighted the [datasets requirement in the doc](https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file0. Hopefully it will better guide the user when they run into this issue :)
