### 2023-05-15 #5313 @brentp
Hi, now I have the problem that there is, as far as I can tell, no way to set the slurm account. I have modified `configs/cluster_config.json` to include:
```json
    "__default__": {
        "ntasks": 1,
        "partition": "quinlan-shared-rw",
        "cpus-per-task": "{threads}",
        "account": "quinlan-rw",
        "mem-per-cpu": "8G",
        "jobname": "QuaC.{rule}.{jobid}",
        "output": "{RULE_LOGS_PATH}/{rule}-%j.log"
    },

```
so a default parition and account is specified. However, the account is not propagated to the snakemake command that looks like this:

```
snakemake \
	--snakefile '/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/workflow/Snakefile' \
	--config project_name='test_project' projects_path='/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/.test/ngs-data' ped='/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/.test/configs/no_priorQC/project_2samples.ped' quac_watch_config='/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/configs/quac_watch/exome_quac_watch_config.yaml' workflow_config='/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/configs/workflow.yaml' unique_id='b27284c4-bf26-43df-9e5d-a05ac4001623' out_dir='/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis' log_dir='/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs' exome='True' include_prior_qc_data='False' allow_sample_renaming='False' \
	--restart-times 1 \
	--use-singularity \
	--singularity-args '--cleanenv --bind /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/tmp:/tmp --bind /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/.test/configs/no_priorQC,/tmp/quac/results/test_project_2samples_exome-no_priorQC/analysis,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/configs/quac_watch/exome_quac_watch_config.yaml,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/verifyBamID,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/reference_genome,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/verifyBamID/exome,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/.test/ngs-data/test_project/analysis,/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/external/dependency_datasets/somalier' \
	--profile '/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/src/slurm/slurm_profile' \
	--cluster-config '/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/configs/cluster_config.json' \
	--cluster 'sbatch --ntasks {cluster.ntasks} --partition {cluster.partition} --cpus-per-task {cluster.cpus-per-task} --mem-per-cpu {cluster.mem-per-cpu} --job-name {cluster.jobname} --output {cluster.output} --parsable'

```
