### 2023-05-17 2cf7940
removes hardcoding in sbatch command; uses cluster config to create it instead
