### 2023-05-18 4696caf
adds sys testing basic ver (based on https://stackoverflow.com/a/72735253/3998252)
