### 2023-05-19 29f3720
sets up dependencies for quac
