### 2023-05-19 a0ff724
removes dryrun option; time to run sys testing
