### 2023-05-19 91894f5
updates doc on cluster config
