### 2023-05-19 #5313 @ManavalanG
@brentp This issue was caused by hardcoded slurm's `sbatch` args in the wrapper script. We have now refactored it to construct the `sbatch` command based on args supplied in the cluster config file (uab-cgds-worthey/quac/pull/73). We expect this would resolve the issue. Sorry about the trouble!

We have also [updated the docs](https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-cluster-config-file) to include cluster config step as part of the requirements.
