### 2023-05-19 3484645
adds try-except to job submission step
