+++ src/run_quac.py
@@ -231,7 +231,12 @@ def main(args):
         "resources": slurm_resources,
     }
 
-    submit_slurm_job(pipeline_cmd, job_dict)
+    try:
+        submit_slurm_job(pipeline_cmd, job_dict)
+    except Exception as e:
+        print ("Job submission step ran into an error:")
+        print(repr(e))
+        raise SystemExit(1)
 
     return None
 
