+++ .github/workflows/system_testing.yml
@@ -14,9 +14,6 @@ jobs:
       run:
         shell: bash -l {0}
         
-    # strategy:
-    #   fail-fast: false
-
     steps:
       - name: checkout repository
         uses: actions/checkout@v2
