### 2023-05-30 bc76168
removes try-except
