### 2023-05-30 0a523b1
pins OS ver; cleans up messages; cleans up commands
