+++ .github/workflows/system_testing.yml
@@ -3,8 +3,8 @@ on: [push, pull_request, workflow_dispatch]
 
 jobs:
   create-env:
-    name: ubuntu-latest
-    runs-on: ubuntu-latest
+    name: ubuntu-20.04
+    runs-on: ubuntu-20.04
     defaults:
       run:
         shell: bash -l {0}
@@ -13,7 +13,7 @@ jobs:
       - name: checkout repository
         uses: actions/checkout@v2
         
-      - name: create environment with mamba
+      - name: Create quac environment
         uses: conda-incubator/setup-miniconda@v2
         with:
           mamba-version: "*"
@@ -22,11 +22,11 @@ jobs:
           activate-environment: quac
           environment-file: configs/env/quac.yaml
       
-      - name: check solution
+      - name: Check conda solution
         run: |
           mamba env export
       
-      - name: test snakemake
+      - name: Check snakemake exists in conda env
         run: |
           which snakemake
           snakemake --version
@@ -34,19 +34,16 @@ jobs:
       - uses: eWaterCycle/setup-singularity@v7
         with:
           singularity-version: 3.8.3
-      - name: Run a singularity container
+      - name: Check singularity is working
         run: |
-          # singularity run docker://alpine cat /etc/os-release
           singularity  --version
 
-      - name: Set up dependecies for QuaC
+      - name: Set up dependencies for QuaC
         run: |
           bash src/setup_dependency_datasets.sh
       
       - name: Run QuaC system testing
         run: |
-          which python
-          ls
           python src/run_quac.py \
                 --project_name test_project \
                 --projects_path ".test/ngs-data/" \
