### 2023-05-30 0e55b96
runs 4 variations of system testing serially
