+++ .github/workflows/system_testing.yml
@@ -15,6 +15,12 @@ on:
       - 'configs/**'
       - 'src/**'
       - 'workflow/**'
+    types:
+      - opened
+      - synchronize
+      - reopened
+      - ready_for_review
+      - converted_to_draft
   workflow_dispatch:
 
 
