+++ docs/Changelog.md
@@ -12,6 +12,10 @@ YYYY-MM-DD  John Doe
 ```
 ---
 
+2023-05-31  Manavalan Gajapathy
+
+* Adds system testing as github actions workflow
+
 2023-05-09  Manavalan Gajapathy
 
 * Adds a verification step in the CLI wrapper script to check if the file/dirpaths to be mounted to singularity already exist as expected (#71)
