### 2023-06-06 #5313 @brentp
I continue to try this and continue to have troubles. Now I get an error:
```
Traceback (most recent call last):
  File "/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/src/run_quac.py", line 437, in <module>
    main(ARGS)
  File "/scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/src/run_quac.py", line 239, in main
    "time": slurm_partition_times[args.slurm_partition],
KeyError: 'quinlan-shared-rw'
```
So I printed out `slurm_partition_times` before that error and see:
```
{'express': '02:00:00', 'short': '12:00:00', 'medium': '50:00:00', 'long': '150:00:00'}
```
then I try to find where to modify that:
```
$ git grep express
LICENSE:  You may not propagate or modify a covered work except as expressly
LICENSE:  In the following three paragraphs, a "patent license" is any express
LICENSE:(such as an express permission to practice a patent or covenant not to
configs/workflow.yaml:  express: "02:00:00"
docs/visualize_pipeline.md:srun --ntasks=1 --cpus-per-task=1 --mem-per-cpu=4096 --partition=express --pty /bin/bash
```
so I modify configs/workflow.yaml to include a time-limit for my partition.
Then I see:
```
sbatch: error: Batch job submission failed: Invalid account or account/partition combination specified
Traceback (most recent call last):
```
So I see in the script that it's not specifying the account only the partition. Even though I have this in `configs/cluster_config.json`
```
    "__default__": {
        "ntasks": 1,
        "partition": "quinlan-shared-rw",
        "cpus-per-task": "{threads}",
        "account": "quinlan-rw",
        "mem-per-cpu": "8G",
        "jobname": "QuaC.{rule}.{jobid}",
        "output": "{RULE_LOGS_PATH}/{rule}-%j.log"
    },
```

The header in the .sh file contains no --account:

```

#SBATCH -e /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs/quac-2023-06-06T09.34.54.561465.%J.err
#SBATCH -o /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs/quac-2023-06-06T09.34.54.561465.%J.out
#SBATCH -J quac-2023-06-06T09.34.54.561465

#SBATCH --partition=quinlan-shared-rw
#SBATCH --ntasks=1
#SBATCH --time=120:00:00
#SBATCH --cpus-per-task=1
#SBATCH --mem-per-cpu=8G
```

It seems it should be present from this PR: https://github.com/uab-cgds-worthey/quac/pull/73/files
and I am on commit: 91894f5 in the `joss_manuscript` branch.

I tried adding:
```
--cluster_config configs/cluster_config.json
```
to the `run_quac.py` args but still account is not propagated.
I also tried removing `--slurm_partition` from the `run_quac.py` args, but then it doesn't grab anything from `cluster_config.json` which results in this header for the sbatch script:
```
#SBATCH -e /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs/quac-2023-06-06T09.37.58.214455.%J.err
#SBATCH -o /scratch/ucgd/lustre-work/quinlan/u6000771/src/quac/data/quac/logs/quac-2023-06-06T09.37.58.214455.%J.out
#SBATCH -J quac-2023-06-06T09.37.58.214455

#SBATCH --partition=short
#SBATCH --ntasks=1
#SBATCH --time=12:00:00
#SBATCH --cpus-per-task=1
#SBATCH --mem-per-cpu=8G
```
So it has ignored the cluster_config.json.

Again, I request that you have someone test this on a different cluster outside your lab.
Note that they must do everything themselves, including finding the `joss_manuscript` branch, modifying the cluster_config, etc. without prompting.

Likely I am doing something very simple that I shouldn't, but I don't know what that is and have tried to find a solution without success.
