### 2023-06-07 uab-cgds-worthey/quac#76 @ManavalanG
CLI script not accounting for user's environment specific slurm args requirements
A reviewer [ran into error](https://github.com/openjournals/joss-reviews/issues/5313#issuecomment-1579010464) due to slurm related issues when slurmpy attempts to submit the parent quac job to slurm.

We believe we fixed a variation of this issue in #73 where snakemake would be submitting jobs to slurm. Ideally we should have fixed above issues as part of that PR, but we didn't catch it back then.

Planned solution:

* When CLI script gathers slurm args to use with parent quac job submission to slurm:

```py
    slurm_resources = {
        "partition": args.slurm_partition,
        "ntasks": "1",
        "time": slurm_partition_times[args.slurm_partition],
        "cpus-per-task": "1" if args.subtasks_slurm else "4",
        "mem-per-cpu": "8G",
    }
```

add a step to read cluster config file and include any args that are not already in `slurm_resources`

* Remove default for `--cluster_config` and force user to provide it manually. Its help message will instruct the user to use `configs/cluster_config.json` as template and edit to match their slurm environment. We think this will let user take a second and set up the configs as needed.
* Update docs to include example on how to edit cluster config to match user's slurm env.
