### 2023-06-07 uab-cgds-worthey/quac#77 @ManavalanG
Have users try out QuaC in Slurm cluster which is not cheaha.
As per [reviewer recommendations](https://github.com/openjournals/joss-reviews/issues/5313#issuecomment-1579010464), we would like to have users try out QuaC in a slurm cluster, which is different to where QuaC is developed and tested (Cheaha).
