### 2023-06-08 0e3e68a
refactors parent slurm job to use user specific sbatch args from cluster
