### 2023-06-08 9877e46
refactors to accept sbatch configs for parent slurm job in a config file
