+++ configs/workflow.yaml
@@ -25,10 +25,3 @@ resources:
   verifybamid:
     no_cpu: 1 # CGDS uses 4
 
-#### slurm partitions ####
-slurm_partitions:
-  # modify below to provide partitions available in your slurm and their time limits
-  express: "02:00:00"
-  short: "12:00:00"
-  medium: "50:00:00"
-  long: "150:00:00"
