### 2023-06-08 828f51d
removes defaults for cluster config files
