### 2023-06-08 fb32306
spits error if cluster config not provided
