### 2023-06-12 514dc02
spits error if cli cluster config expected but not found
