### 2023-06-14 da3f9bb
adds slurm-based system testing back to PR template
