+++ .github/pull_request_template.md
@@ -8,6 +8,7 @@ Please fill in the checklist below and comment as needed.
 - [ ] Is this a feature addition? Briefly describe.
 - [ ] Did you modify QuaC-Watch config file? If so, did you modify multiqc template
   `configs/multiqc_config_template.jinja2` and script `src/quac_watch/create_mutliqc_configs.py` as necessary?
+- [ ] Did you perform system-level testing manually, using `--subtasks_slurm` flag, as described in the [documentation](https://quac.readthedocs.io/en/stable/system_testing/)? Did it pass completely? If not why?
 - [ ] Updated `Changelog.md` file with change logs in recommended format?
 
 
