### 2023-06-14 a986d94
adds sys testing badge to readme
