### 2023-06-15 e754213
adds debugging step for storage
