### 2023-06-15 8756a52
trying try-catch in bash
