+++ .github/workflows/system_testing.yml
@@ -90,8 +90,8 @@ jobs:
                 --outdir "data/quac/results/test_${PROJECT_CONFIG}_wgs-${PRIOR_QC_STATUS}/analysis" \
                 --quac_watch_config "configs/quac_watch/wgs_quac_watch_config.yaml" \
                 --workflow_config "configs/workflow.yaml" \
-                $USE_SLURM
-          df -h
+                $USE_SLURM    || df -h
+          
 
       - name: Run QuaC system testing - Exome mode AND no prior QC data
         run: |
@@ -107,8 +107,7 @@ jobs:
                 --quac_watch_config "configs/quac_watch/exome_quac_watch_config.yaml" \
                 --workflow_config "configs/workflow.yaml" \
                 --exome \
-                $USE_SLURM
-          df -h
+                $USE_SLURM    || df -h
 
       - name: Run QuaC system testing - WGS mode AND uses prior QC data
         run: |
@@ -125,8 +124,7 @@ jobs:
                 --include_prior_qc \
                 --allow_sample_renaming \
                 --workflow_config "configs/workflow.yaml" \
-                $USE_SLURM
-          df -h
+                $USE_SLURM    || df -h
 
       - name: Run QuaC system testing - Exome mode AND uses prior QC data
         run: |
@@ -144,5 +142,4 @@ jobs:
                 --include_prior_qc \
                 --allow_sample_renaming \
                 --workflow_config "configs/workflow.yaml" \
-                $USE_SLURM
-          df -h
+                $USE_SLURM    || df -h
