### 2023-06-20 ce4a420
adds step to free disk space
