### 2023-06-20 0c23956
removes storage calc steps
