### 2023-06-21 0e0ceec
turns off sync PR trigger to avoid duplicate CI trigger - #79
