### 2023-06-21 3ca7487
updates cli options in doc
