### 2023-06-21 a63b76e
removes outdated items in doc
