+++ docs/reqts_configs.md
@@ -11,19 +11,22 @@ In order to run the QuaC pipeline, user needs to
 
 - Linux OS. QuaC was developed and tested in Red Hat Enterprise Linux  v7.9
 
-- [Singularity](https://apptainer.org/) is required but not provided. QuaC pipeline was developed and tested using Singularity v3.5.2.
+- [Singularity](https://apptainer.org/) is required but not provided. QuaC pipeline was developed and tested using
+  Singularity v3.5.2.
 
 !!! note "Cheaha users only"
 
     Singularity is available as module in Cheaha - `Singularity/3.5.2-GCC-5.4.0-2.26`
 
-- Following dependencies are installed as part of the `quac` conda environment. See [installation](./installation.md) for info on creating this conda environment.
+- Following dependencies are installed as part of the `quac` conda environment. See [installation](./installation.md)
+  for info on creating this conda environment.
     - Snakemake-minimal v6.0.5
     - Python v3.6.13
     - Slurmpy v0.0.8
 
 
-- Tools below are used in the QuaC pipeline, and Snakemake automatically installs them as needed during QuaC execution. Therefore, they don't need to be manually installed. For tool versions used, refer to the Snakemake rules.
+- Tools below are used in the QuaC pipeline, and Snakemake automatically installs them as needed during QuaC execution.
+  Therefore, they don't need to be manually installed. For tool versions used, refer to the Snakemake rules.
     - qualimap
     - picard
     - mosdepth
@@ -37,26 +40,37 @@ In order to run the QuaC pipeline, user needs to
 
 ## Set up workflow config file
 
-QuaC requires a workflow config file in yaml format (default: `configs/workflow.yaml`). It provides following info to QuaC: 
+QuaC requires a workflow config file in yaml format (default: `configs/workflow.yaml`). It provides following info to
+QuaC: 
 
 - Filepaths to necessary dataset dependencies required by certain QC tools used in QuaC.
     - **Use script `src/setup_dependency_datasets.sh` to download and set up these datasets.**
-    - To run this script, use command: `bash src/setup_dependency_datasets.sh`. Output will be saved at `data/external/dependency_datasets`.
+    - To run this script, use command: `bash src/setup_dependency_datasets.sh`. Output will be saved at
+      `data/external/dependency_datasets`.
 - Hardware resource configs
 
-!!! tip
+!!! tip 
+
     Custom workflow config file can be provided to QuaC via `--workflow_config`.
 
 
 ## Set up cluster config file
 
-QuaC requires a cluster config file (default: `configs/snakemake_cluster_config.json`), where slurm resources to be utilized by QuaC are configured. You may want to modify `__default__` section in this file to suit your slurm system. For example, if your slurm requires `--account` to be supplied to `sbatch`, you will need to add it to the `__default__` (e.g., "account": "uname").
+QuaC requires a cluster config file (default: `configs/snakemake_cluster_config.json`), where slurm resources to be
+utilized by QuaC are configured. You may want to modify `__default__` section in this file to suit your slurm system.
+For example, if your slurm requires `--account` to be supplied to `sbatch`, you will need to add it to the `__default__`
+(e.g., "account": "uname").
 
-!!! tip
+!!! tip 
+
     Custom cluster config file can be provided to QuaC via `--cluster_config`.
 
 
 ## Pull Singularity images [Optional]
 
[…]
