### 2023-06-21 c8a6162
reads slurm resources only when necessary
