### 2023-06-21 eeafdc3
removes "or"  conditional as it ends up passing sys test when it is supposed to be a failure - https://github.com/uab-cgds-worthey/quac/actions/runs/5338057275
