### 2023-06-21 68da0c1
checks tool dependencies exist
