### 2023-06-22 db4c29d
removes --snakemake_slurm as it is redundant
