### 2023-06-22 a3e7c8c
removes redundant --subtasks_slurm option
