### 2023-06-22 8961123
singularity compiling fails in ubuntu 22; falling back to older version
