+++ .github/workflows/system_testing.yml
@@ -25,7 +25,7 @@ on:
 jobs:
   system-testing:
     name: System testing - QuaC
-    runs-on: ubuntu-22.04
+    runs-on: ubuntu-20.04
     defaults:
       run:
         shell: bash -l {0}
