### 2023-06-26 61cc238
updates doc about removal of subtasks_slurm and snakemake_slurm
