+++ docs/quac_cli.md
@@ -13,8 +13,7 @@ usage: run_quac.py [-h] [--project_name] [--projects_path] [--pedigree]
                    [--quac_watch_config] [--workflow_config]
                    [--snakemake_cluster_config] [--outdir] [--tmp_dir]
                    [--exome] [--include_prior_qc] [--allow_sample_renaming]
-                   [--subtasks_slurm] [-e] [-n] [--cli_cluster_config]
-                   [--log_dir] [--snakemake_slurm]
+                   [-e] [-n] [--cli_cluster_config] [--log_dir]
 
 Command line interface to QuaC pipeline.
 
@@ -31,7 +30,8 @@ QuaC workflow options:
                         Watch. See directory 'configs/quac_watch/' in quac
                         repo for the included config files. (default: None)
   --workflow_config     YAML config path specifying filepath to dependencies
-                        of tools used in QuaC (default: configs/workflow.yaml)
+                        of QC tools used in snakemake workflow (default:
+                        configs/workflow.yaml)
   --snakemake_cluster_config
                         Cluster config json file. Needed for snakemake to run
                         jobs in cluster. Edit template file
@@ -49,10 +49,6 @@ QuaC workflow options:
   --allow_sample_renaming
                         Flag to allow sample renaming in MultiQC report. See
                         documentation for more info. (default: False)
-  --subtasks_slurm      Flag indicating that the main Snakemake process of
-                        QuaC should submit subtasks of the workflow as Slurm
-                        jobs instead of running them on the same machine as
-                        itself (default: False)
   -e , --extra_args     Pass additional custom args to snakemake. Equal symbol
                         is needed for assignment as in this example: -e='--
                         forceall' (default: None)
@@ -68,11 +64,6 @@ QuaC wrapper options:
                         environment. (default: None)
   --log_dir             Directory path where logs (both workflow's and
                         wrapper's) will be stored (default: data/quac/logs)
-  --snakemake_slurm     Flag indicating that the main Snakemake process of
-                        QuaC should be submitted to run in a Slurm job instead
-                        of executing in the current environment. Useful for
-                        headless execution on Slurm-based HPC systems.
-                        (default: False)
 ```
 
 ### Useful features
@@ -82,15 +73,16 @@ Besides the basic features, wrapper script [`src/run_quac.py`](../src/run_quac.p
 - Pass custom snakemake args using option `--extra_args`.
 - Dry-run snakemake using flag `--dryrun`. Note that this is same as `--extra_args='-n'`.
 - Override cluster config file passed to snakemake using `--snakemake_cluster_config`.
-- Submit snakemake process to Slurm, instead of running it locally, using `--snakemake_slurm`. 
-- Submit jobs triggered by snakemake workflow to Slurm using `--subtasks_slurm`.
+- Submit snakemake process to Slurm, instead of running it locally, using `--cli_cluster_config`. 
+- Submit jobs triggered by snakemake workflow to Slurm using `--snakemake_cluster_config`.
 
 ## Minimal example
 
-Minimal example to run the wrapper script, which in turn will execute the QuaC pipeline:
+Minimal example to run the wrapper script, which in turn will execute the QuaC pipeline locally:
 
 ```sh
-# For Cheaha users only. Set up environment. 
+# First set up dependencies in the environment. 
+### Cheaha users can set them up as follows. 
 module reset
 module load Anaconda3/2020.02
 module load Singularity/3.5.2-GCC-5.4.0-2.26
@@ -101,8 +93,9 @@ conda activate quac
 # run CLI/wrapper script
 python src/run_quac.py \
       --project_name "PROJECT_DUCK" \
[…]
