+++ docs/system_testing.md
@@ -32,7 +32,8 @@ conda activate quac
 ## use slurm or not
 # use this to submit jobs to slurm for the parent snakemake process 
 # as well as for the snakemake triggered jobs
-USE_SLURM="--snakemake_slurm --subtasks_slurm"
+USE_SLURM="--cli_cluster_config configs/cli_cluster_config.json 
+           --snakemake_cluster_config configs/snakemake_cluster_config.json"
 # USE_SLURM=""  # uncomment this, comment out the above line, and use this if you don't want to use slurm at all
 
 
