+++ configs/workflow.yaml
@@ -15,7 +15,7 @@ datasets:
     # retrieved from https://github.com/Griffan/VerifyBamID and 'chr' prefix added to the contigs
     svd_dat_exome: "data/external/dependency_datasets/verifyBamID/exome/1000g.phase3.10k.b38_chr.exome.vcf.gz.dat"
 
-#### hardware resources ####
+#### hardware resources for certain tools used in snakemake piepline ####
 resources:
   qualimap_bamqc:
     no_cpu: 1 # CGDS uses 2
