### 2023-06-26 b38b618
adds doc on SLURM config modification
