+++ docs/reqts_configs.md
@@ -3,8 +3,8 @@
 In order to run the QuaC pipeline, user needs to
 
 1. [Install the pipeline and set up the conda environment](./installation.md).
-2. Set up config files specifying paths and hardware resources required by QC tools used in the pipeline.
-3. Set up cluster config file to suit your Slurm setup.
+2. Set up workflow config files specifying paths and hardware resources required by QC tools used in the pipeline.
+3. Set up cluster config files if you will be submitting jobs to SLURM cluster.
 4. (Optional) Run QuaC pipeline just to create singularity environments using the system testing datasets.
 
 ## Requirements
@@ -18,12 +18,14 @@ In order to run the QuaC pipeline, user needs to
 
     Singularity is available as module in Cheaha - `Singularity/3.5.2-GCC-5.4.0-2.26`
 
+- [SLURM](https://slurm.schedmd.com/)
+    - Optional. Needed only if you will be supplying `--snakemake_slurm` to `src/run_quac.py`.
+
 - Following dependencies are installed as part of the `quac` conda environment. See [installation](./installation.md)
   for info on creating this conda environment.
     - Snakemake-minimal v6.0.5
     - Python v3.6.13
     - Slurmpy v0.0.8
-
 
 - Tools below are used in the QuaC pipeline, and Snakemake automatically installs them as needed during QuaC execution.
   Therefore, they don't need to be manually installed. For tool versions used, refer to the Snakemake rules.
@@ -47,24 +49,117 @@ QuaC:
     - **Use script `src/setup_dependency_datasets.sh` to download and set up these datasets.**
     - To run this script, use command: `bash src/setup_dependency_datasets.sh`. Output will be saved at
       `data/external/dependency_datasets`.
-- Hardware resource configs
+- Hardware resource configs for certain QC tools used in snakemake piepline
 
 !!! tip 
 
     Custom workflow config file can be provided to QuaC via `--workflow_config`.
 
 
-## Set up cluster config file
+## Set up SLURM cluster config file
 
-QuaC requires a cluster config file (default: `configs/snakemake_cluster_config.json`), where slurm resources to be
-utilized by QuaC are configured. You may want to modify `__default__` section in this file to suit your slurm system.
-For example, if your slurm requires `--account` to be supplied to `sbatch`, you will need to add it to the `__default__`
-(e.g., "account": "uname").
+QuaC's wrapper script `src/run_quac.py` allows you to submit jobs to SLURM scheduler at two levels:
 
-!!! tip 
+* submits the script that will run the snakemake workflow as a SLURM job (using `--cli_cluster_config`)
+* allows snakemake to submit jobs to SLURM (using `--snakemake_cluster_config`)
 
-    Custom cluster config file can be provided to QuaC via `--cluster_config`.
+These config files specify the resources to be requested when submitting the job to SLURM.
 
+
+### `--cli_cluster_config` config file
+
+We provide a template file `configs/cli_cluster_config.json` to use with `--cli_cluster_config`. Its contents are shown
+below:
+
+```json
+{!configs/cli_cluster_config.json!}
+```
+
+As SLURM configuration at your cluster environment might be different than ours, you may need to modify this template
+file to suit your SLURM setup.
+
+For example, some SLURM might need `account` info when submitting jobs. So you may modify the config file as:
+
+```json hl_lines="2"
+{
+    "account": "your_account_name",     <----- edited line
+    "partition": "express",
+    "ntasks": "1",
+    "time": "02:00:00",
+    "cpus-per-task": "1",
+    "mem-per-cpu": "8G"
+}
[…]
