+++ docs/Changelog.md
@@ -20,7 +20,8 @@ YYYY-MM-DD  John Doe
   (`src/slurm/slurm_profile/config.yaml`)
 * Removes `--subtasks_slurm option` and `--snakemake_slurm` as they were redundant
 * Checks if tool dependencies are available in user environment.
-
+* Adds `time` resource to `configs/snakemake_cluster_config.json`
+  
 
 2023-05-31  Manavalan Gajapathy
 
