### 2023-06-27 c587709
removes PR trigger from system testing workflow
