+++ .github/workflows/system_testing.yml
@@ -7,19 +7,6 @@ on:
       - "configs/**"
       - "src/**"
       - "workflow/**"
-  pull_request:
-    paths:
-      - ".github/workflows/system_testing.yml"
-      - ".test/**"
-      - "configs/**"
-      - "src/**"
-      - "workflow/**"
-    types:
-      - opened
-      # - synchronize
-      - reopened
-      - ready_for_review
-      - converted_to_draft
   workflow_dispatch:
 
 jobs:
