### 2023-06-28 994699a
fixes usage of deprecated arg
Co-authored-by: Brandon M Wilk <wilkb777@gmail.com>
