### 2023-06-28 405beb0
Update docs/reqts_configs.md
Co-authored-by: Brandon M Wilk <wilkb777@gmail.com>
