### 2023-06-28 c9ad9f5
does dependency check in one step
Co-authored-by: Brandon M Wilk <wilkb777@gmail.com>
