### 2023-06-28 d346c93
refactors dependency checking so can custom print error msg
