+++ src/run_quac.py
@@ -54,8 +54,10 @@ def check_mount_paths_exist(paths):
         raise SystemExit(
             (
                 f"ERROR: Following directories that are part of your input were not found: \n{fail_paths}"
-                "\n\nNOTE: This was likely caused by missing datasets specified in the workflow config file (supplied via --workflow_config)."
-                "\nPlease refer to docs on how to use our script to download the necessary datasets - https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file"
+                "\n\nNOTE: This was likely caused by missing datasets specified in the workflow config file"
+                "(supplied via --workflow_config)."
+                "\nPlease refer to docs on how to use our script to download the necessary datasets -"
+                " https://quac.readthedocs.io/en/latest/reqts_configs/#set-up-workflow-config-file"
             )
         )
 
