### 2023-06-28 21026f2
adds test to check singularity is working properly
