### 2023-06-28 7a23236
updates error msg
