### 2023-06-30 1bd842d
restructures docs to streamline install and config
