+++ docs/reqts_configs.md
@@ -11,15 +11,7 @@ In order to run the QuaC pipeline, user needs to
 
 - Linux OS. QuaC was developed and tested in Red Hat Enterprise Linux  v7.9
 
-- [Singularity](https://apptainer.org/) is required but not provided. QuaC pipeline was developed and tested using
-  Singularity v3.5.2.
 
-!!! note "Cheaha users only"
-
-    Singularity is available as module in Cheaha - `Singularity/3.5.2-GCC-5.4.0-2.26`
-
-- [SLURM](https://slurm.schedmd.com/)
-    - Optional. Needed only if you will be supplying `--snakemake_cluster_config` and/or `--cli_cluster_config` to `src/run_quac.py`.
 
 - Following dependencies are installed as part of the `quac` conda environment. See [installation](./installation.md)
   for info on creating this conda environment.
@@ -40,126 +32,12 @@ In order to run the QuaC pipeline, user needs to
     - multiqc
 
 
-## Set up workflow config file
-
-QuaC requires a workflow config file in yaml format (default: `configs/workflow.yaml`). It provides following info to
-QuaC: 
-
-- Filepaths to necessary dataset dependencies required by certain QC tools used in QuaC.
-    - **Use script `src/setup_dependency_datasets.sh` to download and set up these datasets.**
-    - To run this script, use command: `bash src/setup_dependency_datasets.sh`. Output will be saved at
-      `data/external/dependency_datasets`.
-- Hardware resource configs for certain QC tools used in snakemake piepline
-
-!!! tip 
-
-    Custom workflow config file can be provided to QuaC via `--workflow_config`.
 
 
 ## Set up SLURM cluster config file
 
-QuaC's wrapper script `src/run_quac.py` allows you to submit jobs to SLURM scheduler at two levels:
-
-* using `--cli_cluster_config` - submits the script that will run the snakemake workflow as a SLURM job
-* using `--snakemake_cluster_config` - allows snakemake to submit jobs to SLURM
-
-These config files specify the resources to be requested when submitting the job to SLURM.
 
 
-### a. Set up `--cli_cluster_config` config file
-
-We provide a template file `configs/cli_cluster_config.json` to use with `--cli_cluster_config`. Its contents are shown
-below:
-
-```json
-{!configs/cli_cluster_config.json!}
-```
-
-As SLURM configuration in your cluster environment likely differs from other HPC clusters, you may need to modify this template
-file to suit your SLURM setup.
-
-For example, some SLURM might need `account` info when submitting jobs. So you may modify the config file, assuming
-other args are acceptable, as:
-
-```json hl_lines="2"
-{
-    "account": "your_account_name",     <----- edited line
-    "partition": "express",
-    "ntasks": "1",
-    "time": "02:00:00",
-    "cpus-per-task": "1",
-    "mem-per-cpu": "8G"
-}
-```
-
-If your SLURM requires `qos` and doesn't use `partition`, config file can be modified as:
-
-```json
-{
-    "qos": "qos_tag",           <----- edited line
-    // "partition": "express",  <----- edited line
-    "ntasks": "1",
-    "time": "02:00:00",
-    "cpus-per-task": "1",
-    "mem-per-cpu": "8G"
-}
-```
-
-### b. Set up `--snakemake_cluster_config` config file
-
-We provide a template file `configs/snakemake_cluster_config.json` to use with `--snakemake_cluster_config`. Its
-contents are shown below:
-
-```json
-{!configs/snakemake_cluster_config.json!}
-```
-
[…]
