+++ docs/installation_configuration.md
@@ -1,0 +1,196 @@
+# Installation & Configuration
+
+This involves the following steps:
+
+* Fetch the source code for QuaC
+* Set up necessary configurations
+* Create the conda environment `quac`
+* Test run to check QuaC runs as expected
+
+## Requirements
+
+Requirements below are necessary to install and run QuaC.
+
+- Linux OS. Tested in Red Hat Enterprise Linux v7.9
+- Git
+- Anaconda/miniconda. Tested using Anaconda3/2020.02
+- [Singularity](https://apptainer.org/) is required but not provided. Tested using Singularity v3.5.2.
+- [SLURM](https://slurm.schedmd.com/)
+    - Needed if you will be supplying `--snakemake_cluster_config` and/or `--cli_cluster_config` to `src/run_quac.py`.
+
+!!! note "Cheaha users only"
+
+    * Conda is Available as module in Cheaha - `Anaconda3/2020.02`
+    * Singularity is available as module in Cheaha - `Singularity/3.5.2-GCC-5.4.0-2.26`
+
+
+## Installation
+
+Installation involves
+
+- fetching the source code
+- creating the conda environment
+
+
+### Retrieve source code
+
+Go to the directory of your choice and run the command below.
+
+```sh
+git clone https://github.com/uab-cgds-worthey/quac.git
+```
+
+
+### Create conda environment
+
+Conda environment will install all necessary dependencies, including snakemake, to run the QuaC workflow.
+
+```sh
+cd /path/to/quac/repo
+
+# For use only at Cheaha in UAB. Load conda into environment.
+module reset
+module load Anaconda3/2020.02
+
+# create conda environment. Needed only the first time.
+conda env create --file configs/env/quac.yaml
+
+# activate conda environment
+conda activate quac
+
+# if you need to update the existing environment
+conda env update --file configs/env/quac.yaml
+```
+
+
+## Configuration
+
+### Workflow config 
+
+QuaC requires a workflow config file in yaml format (default: `configs/workflow.yaml`). It provides following info to
+QuaC: 
+
+- Filepaths to necessary dataset dependencies required by certain QC tools used in QuaC.
+    - ** We provide `src/setup_dependency_datasets.sh` to download and set up these datasets.**
+    - To run this script, change into QuaC directory and use command: `bash src/setup_dependency_datasets.sh`. Output will be saved at
+      `data/external/dependency_datasets`.
+- Hardware resource configs for certain QC tools used in snakemake piepline
+
+!!! tip 
+
+    Custom workflow config file can be provided to QuaC via `--workflow_config`.
+
+### SLURM config
+
+QuaC's wrapper script `src/run_quac.py` allows you to submit jobs to SLURM scheduler at two levels:
+
+* using `--cli_cluster_config` - submits the script that will run the snakemake workflow as a SLURM job
+* using `--snakemake_cluster_config` - allows snakemake to submit jobs to SLURM
+
+These config files specify the resources to be requested when submitting the job to SLURM.
+
+### a. Set up `--cli_cluster_config` config file
+
+We provide a template file `configs/cli_cluster_config.json` to use with `--cli_cluster_config`. Its contents are shown
+below:
+
+```json
+{!configs/cli_cluster_config.json!}
+```
+
+As SLURM configuration in your cluster environment likely differs from other HPC clusters, you may need to modify this
+template file to suit your SLURM setup.
+
+For example, some SLURM might need `account` info when submitting jobs. So you may modify the config file, assuming
+other args are acceptable, as:
+
+```json hl_lines="2"
+{
[…]
