### 2023-06-30 8578598
moves down conda env creation
