+++ docs/installation_configuration.md
@@ -5,7 +5,7 @@ This involves the following steps:
 * Fetch the source code for QuaC
 * Set up necessary configurations
 * Create the conda environment `quac`
-* Test run to check QuaC runs as expected
+* Test run to check QuaC works as expected
 
 ## Requirements
 
@@ -22,7 +22,6 @@ Requirements below are necessary to install and run QuaC.
 
     * Conda is Available as module in Cheaha - `Anaconda3/2020.02`
     * Singularity is available as module in Cheaha - `Singularity/3.5.2-GCC-5.4.0-2.26`
-
 
 
 ## Retrieve QuaC source code
@@ -184,3 +183,41 @@ conda activate quac
 # if you need to update the existing environment
 conda env update --file configs/env/quac.yaml
 ```
+
+## Test run QuaC
+
+When you have completed all the installation and configuration steps described so far, it is time to test run QuaC to check it runs as expected using the commands below.  
+
+```sh
+# Be sure conda and singularity are available in your enviroment
+# Below snippet is for Cheaha users only.
+module reset
+module load Anaconda3/2020.02
+module load Singularity/3.5.2-GCC-5.4.0-2.26
+
+# activate conda env
+conda activate quac
+
+# specify slurm config files to use
+USE_SLURM="--cli_cluster_config configs/cli_cluster_config.json 
+           --snakemake_cluster_config configs/snakemake_cluster_config.json"
+
+PROJECT_CONFIG="project_2samples"
+PRIOR_QC_STATUS="no_priorQC"
+
+python src/run_quac.py \
+      --project_name test_project \
+      --projects_path ".test/ngs-data/" \
+      --pedigree ".test/configs/${PRIOR_QC_STATUS}/${PROJECT_CONFIG}.ped" \
+      --outdir "data/quac/results/test_${PROJECT_CONFIG}_wgs-${PRIOR_QC_STATUS}/analysis" \
+      --quac_watch_config "configs/quac_watch/wgs_quac_watch_config.yaml" \
+      --workflow_config "configs/workflow.yaml" \
+      $USE_SLURM
+```
+
+!!! info
+
+    In case you run into an error, please double check that configuration files were properly set up.
+
+Note that the first time you run the snakemake workflow, it will first retrieve the necessary singularity images and then run the QC jobs. This process may take about 30mins to 1hr to complete the job. If it takes way longer than an hour, please check the log files (specified using `--log_dir`) and your SLURM jobs to identify the cause.
+
