### 2023-07-03 uab-cgds-worthey/quac#76 @ManavalanG
Quick note - solution we ended up using was to provide slurm info as a input config file.
