### 2023-07-13 16a52c5
removes broken links to files outside docs dir
