+++ docs/system_testing.md
@@ -1,10 +1,9 @@
 # System testing
 
 The system testing implemented for this pipeline tests whether the pipeline runs from start to finish without any error.
-This testing uses test datasets present in [`.test/ngs-data/test_project`](../.test/ngs-data/test_project), which
-reflects a test project containing four samples -- Two samples without priorQC data (`no_priorQC`) and two with priorQC
-data (`include_priorQC`). [See .test/README.md](../.test/README.md) for more info on how these test datasets were
-created.
+This testing uses test datasets present in `.test/ngs-data/test_project`, which reflects a test project containing four
+samples -- Two samples without priorQC data (`no_priorQC`) and two with priorQC data (`include_priorQC`). See
+`.test/README.md` for more info on how these test datasets were created.
 
 !!! warning
 
