+++ mkdocs.yaml
@@ -7,6 +7,7 @@ nav:
   - QuaC-Watch: quac_watch.md
   - System testing: system_testing.md
   - Pipeline visualization: visualize_pipeline.md
+  - Frequently Asked Questions: faq.md
   - Changelog: Changelog.md
   - QC review:
       - Sample QC review system: sample_qc_review_system.md
