### 2023-07-17 #5313 @ManavalanG
@brentp @Redmar-van-den-Berg We have modified how slurm configs are provided to QuaC
(https://github.com/uab-cgds-worthey/quac/pull/80), and we also made several other changes to improve user experience (checking dependencies exist, singularity works as expected, etc.). Also, as per @brentp's recommendation, we had three users install quac, configure it, run system testing on another Slurm HPC environment and provide feedback. The configuration updates as well as updates to the documentation are based on their feedback (https://github.com/uab-cgds-worthey/quac/pull/82). Please let me know if you have any questions.
