### 2023-08-21 #5313 @brentp
Hi, sorry for the delay. I have just moved continents and finally now set up to try this.

I was able to get this running. I had some trouble with setup dependencies as my gzip gave an error an the trailing "garbage" from the razip/bgzip'ed fasta. I changed it like this and it works

```patch
diff --git a/src/setup_dependency_datasets.sh b/src/setup_dependency_datasets.sh
index 0208e62..2f5f22f 100755
--- a/src/setup_dependency_datasets.sh
+++ b/src/setup_dependency_datasets.sh
@@ -16,7 +16,7 @@ mkdir -p $REF_GENOME_DIR && cd $REF_GENOME_DIR
 REF_GENOME_FNAME="GCA_000001405.15_GRCh38_no_alt_analysis_set.fna.gz"
 REF_GENOME_URL=ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCA/000/001/405/GCA_000001405.15_GRCh38/seqs_for_alignment_pipelines.ucsc_ids/${REF_GENOME_FNAME}
 curl -L $REF_GENOME_URL -o $REF_GENOME_FNAME
-gzip -d $REF_GENOME_FNAME
+gzip -c $REF_GENOME_FNAME > $(basename $REF_GENOME_FNAME .gz)

 ##### retrieve somalier tool dependencies #####
 echo "setting up somalier dependency datasets..."
```

I will update my review shortly.
