### 2023-08-21 dc4bef2
modifies gzip decompressing cmd - https://github.com/openjournals/joss-reviews/issues/5313#issuecomment-1686290957
