### 2023-10-05 uab-cgds-worthey/quac#85 @Redmar-van-den-Berg
Please add documentation on how to set QC limits
**Describe the bug**
The documentation contains an example of how to run the pipeline under the `System testing` section, but it does not explain how to set the QC limits, which are specified in `configs/quac_watch/wgs_quac_watch_config.yaml`. Specifically, which tools are available, which values are available in each tool, which (kinds of) limits can be set, etc.

**Additional context**
Preferably, there should also be some error checking that warns the user when the configuration is not complete. For example, if you remove the `max` key for one of the modules (because you only care about a minimum value), the whole pipeline will crash with a `KeyError`, rather than warning the user.
