### 2023-10-05 uab-cgds-worthey/quac#86 @Redmar-van-den-Berg
Feature request: add support for specifying samples directly
**Describe the bug**
Currently, quac expects a certain directory structure to identify samples and bam/vcf files, which makes it difficult to use with existing pipelines, which may have a different folder structure. It would be nice to have a way to explicitly specify samples and files, so that users do not have to replicate the exact folder structure that quac expects.

**Additional context**
One way to do this would be to add support for something like [PEP](http://pep.databio.org/en/latest/), which can be as simple as a csv file, but also supports more complex project configurations. Snakemake already has native support for PEP, see [here](https://snakemake.readthedocs.io/en/stable/snakefiles/configuration.html#configuring-scientific-experiments-via-peps), so this should not be difficult to add to quac.
