### 2023-10-05 #5313 @Redmar-van-den-Berg
@lpantano Apologies for the delay, I've now completed the checklist.

There are still two open issues:
- A request to improve the user documentation (https://github.com/uab-cgds-worthey/quac/issues/85)
- A feature request to no longer rely on a fixed project structure to identify samples and files (https://github.com/uab-cgds-worthey/quac/issues/86)
