### 2023-10-05 604af7f
refactors to use file inputs from sample config
