+++ src/run_quac.py
@@ -19,6 +19,7 @@ import json
 import yaml
 from singularity_status import test_singularity
 from slurm.submit_slurm_job import submit_slurm_job
+from read_sample_config import read_sample_config
 
 
 def make_dir(d):
@@ -91,8 +92,7 @@ def read_workflow_config(workflow_config_fpath):
 
 
 def gather_mount_paths(
-    projects_path,
-    project_name,
+    sample_config,
     pedigree_path,
     out_dir,
     log_dir,
@@ -105,10 +105,14 @@ def gather_mount_paths(
 
     mount_paths = set()
 
-    # project path
-    project_path = Path(projects_path) / project_name / "analysis"
-    make_dir(project_path)
-    mount_paths.add(project_path)
+    # sample_config
+    mount_paths.add(Path(sample_config).parent)
+
+    # input filepaths from sample config
+    samples_dict = read_sample_config(sample_config)
+    for sample_val in samples_dict.values():
+        for val_fpath in sample_val.values():
+            mount_paths.add(Path(val_fpath).parent)
 
     # pedigree filepath
     mount_paths.add(Path(pedigree_path).parent)
@@ -168,8 +172,6 @@ def create_snakemake_command(args, repo_path, mount_paths):
     tmp_dir = args.tmp_dir
 
     quac_configs = {
-        "project_name": args.project_name,
-        "projects_path": args.projects_path,
         "ped": args.pedigree,
         "quac_watch_config": args.quac_watch_config,
         "workflow_config": args.workflow_config,
@@ -234,8 +236,7 @@ def main(args):
 
     # process user's input-output config file and get singularity bind paths
     mount_paths = gather_mount_paths(
-        args.projects_path,
-        args.project_name,
+        args.sample_config,
         args.pedigree,
         args.outdir,
         args.log_dir,
@@ -270,7 +271,7 @@ def main(args):
         "resources": slurm_resources,
     }
 
-    submit_slurm_job(pipeline_cmd, job_dict)
+    # submit_slurm_job(pipeline_cmd, job_dict)
 
     return None
 
@@ -314,19 +315,14 @@ if __name__ == "__main__":
     WORKFLOW = PARSER.add_argument_group("QuaC snakemake workflow options")
 
     WORKFLOW.add_argument(
-        "--project_name",
-        help="Project name. Required.",
-        required=True,
-    )
-    WORKFLOW.add_argument(
-        "--projects_path",
-        help="Path where all projects are hosted. Do not include project name here. Required.",
-        type=lambda x: is_valid_dir(PARSER, x),
+        "--sample_config",
+        help="Sample config file in TSV format. Provides sample name and necessary input filepaths (bam, vcf, etc.)",
+        type=lambda x: is_valid_file(PARSER, x),
         required=True,
     )
     WORKFLOW.add_argument(
         "--pedigree",
-        help="Pedigree filepath. Must correspond to the project supplied via --project_name. Required.",
+        help="Pedigree filepath. Must correspond to samples mentioned in configfile via --sample_config. Required.",
         type=lambda x: is_valid_file(PARSER, x),
         required=True,
     )
@@ -426,10 +422,6 @@ if __name__ == "__main__":
     if not ARGS.quac_watch_config:
         raise SystemExit(
             "Error. Quac-watch config is missing. Please supply using --quac_watch_config."
-        )
-    if not ARGS.projects_path:
-        raise SystemExit(
-            "Error. 'Projects path' not provided. Please supply using --projects_path."
         )
 
     main(ARGS)
