### 2023-10-05 d998273
updates testing command
