### 2023-10-05 d5a8d5e
updates doc on cli help
