+++ src/read_sample_config.py
@@ -14,7 +14,7 @@ def get_full_path(x):
 
 def is_valid_file(fpath):
     if not Path(fpath).is_file():
-        print(f"File provided in sample config file was not found: {fpath}")
+        print(f"ERROR: File provided in sample config file was not found: {fpath}")
         raise SystemExit(1)
 
     return get_full_path(fpath)
@@ -36,6 +36,16 @@ def read_sample_config(config_f):
                 raise SystemExit(1)
 
             samples_dict[sample] = {"vcf": vcf, "bam": bam}
+
+            for colname in ["capture_bed"]:
+                if colname in row:
+                    samples_dict[sample][colname] = is_valid_file(row[colname])
+
+                    if colname == "capture_bed" and not row["capture_bed"].endswith(".bed"):
+                        print(
+                            f"ERROR: Capture bed filename is required to end with '.bed' extension but it did not: '{row['capture_bed']}'"
+                        )
+                        raise SystemExit(1)
 
     return samples_dict
 
