### 2023-10-05 7bad790
returns header; removes file suffix verification
