### 2023-10-05 96fa05b
reads sample configfile and verifies columns exist
