### 2023-10-05 b37fece
updates command in doc
