+++ docs/system_testing.md
@@ -66,15 +66,15 @@ PRIOR_QC_STATUS="include_priorQC"
 
 # WGS mode
 python src/run_quac.py \
-      --project_name test_project \
-      --projects_path ".test/ngs-data/" \
-      --pedigree ".test/configs/${PRIOR_QC_STATUS}/${PROJECT_CONFIG}.ped" \
+      --sample_config ".test/configs/include_priorQC/sample_config/project_2samples_wgs.tsv" \
+      --pedigree ".test/configs/${PRIOR_QC_STATUS}/pedigree/${PROJECT_CONFIG}.ped" \
       --outdir "data/quac/results/test_${PROJECT_CONFIG}_wgs-${PRIOR_QC_STATUS}/analysis" \
       --quac_watch_config "configs/quac_watch/wgs_quac_watch_config.yaml" \
       --include_prior_qc \
-      --allow_sample_renaming \
       --workflow_config "configs/workflow.yaml" \
       $USE_SLURM
+
+    #   --allow_sample_renaming \
 
 # Exome mode
 python src/run_quac.py \
