### 2023-10-05 c5f882f
adds sample config for exome mode
