### 2023-10-06 9ebb14e
ensures expected columns are present
