+++ src/run_quac.py
@@ -61,9 +61,14 @@ def check_sample_configs(fpath, exome_mode, include_prior_qc, allow_sample_renam
                 )
                 raise SystemExit(1)
 
-    # TODO
-    if include_prior_qc and "TODO" not in header:
-        pass
+    if include_prior_qc:
+        columns = ["fastqc_raw", "fastqc_trimmed", "fastq_screen", "dedup"]
+        missing_columns = list(set(columns).difference(set(header)))
+        if len(missing_columns):
+            print(
+                f"ERROR: Columns missing in sample config file but needed when flag --include_prior_qc is used: {missing_columns}"
+            )
+            raise SystemExit(1)
 
     # TODO
     if allow_sample_renaming and "TODO" not in header:
@@ -314,7 +319,7 @@ def main(args):
         "resources": slurm_resources,
     }
 
-    submit_slurm_job(pipeline_cmd, job_dict)
+    # submit_slurm_job(pipeline_cmd, job_dict)
 
     return None
 
