### 2023-10-06 563d166
updates command in doc
