+++ docs/system_testing.md
@@ -71,9 +71,9 @@ python src/run_quac.py \
       --outdir "data/quac/results/test_${PROJECT_CONFIG}_wgs-${PRIOR_QC_STATUS}/analysis" \
       --quac_watch_config "configs/quac_watch/wgs_quac_watch_config.yaml" \
       --include_prior_qc \
+      --allow_sample_renaming \
       --workflow_config "configs/workflow.yaml" \
       $USE_SLURM
-    #   --allow_sample_renaming \
 
 # Exome mode
 python src/run_quac.py \
@@ -83,9 +83,9 @@ python src/run_quac.py \
       --quac_watch_config "configs/quac_watch/exome_quac_watch_config.yaml" \
       --exome \
       --include_prior_qc \
+      --allow_sample_renaming \
       --workflow_config "configs/workflow.yaml" \
       $USE_SLURM
-    #   --allow_sample_renaming \
 ```
 
 !!! note
