### 2023-10-06 e649368
removes outdated section
