### 2023-10-06 d361c64
adds desc for sample configfile
