### 2023-10-06 cb9605a
removes outdated text
