+++ docs/input_output.md
@@ -45,30 +45,11 @@ supplied in pedigree file will be processed by QuaC and all of these samples mus
     create a dummy pedigree file, which will lack sex (unless project tracking sheet is provided), relatedness and
     affected status info. See header of the script for usage instructions. 
 
-
-*Optionally*, QuaC can also utilize QC results produced by [certain
-tools](./index.md#optional-qc-output-consumed-by-quac) when run with flag `--include_prior_qc`. In this case, following
-directory structure is expected.
-
-
-!!! note "CGDS users only"
-
-    Output (bam, vcf and QC output) produced by CGDS's small variant caller pipeline can be readily used as input to
-    QuaC with flags `--include_prior_qc` and `--allow_sample_renaming`.
-
-### Example project structure
-
-Refer to system testing directory `.test/` in the repo for an example project to see an example project with above
-mentioned directory structure needed as input. In this setup, projects A and B have prior QC data included, whereas
-samples C and D do not have them. Refer to pedigree files under `.test/configs/` on how these example samples were used
-as input to QuaC. 
-
-
 ## Output
 
 QuaC results are stored at the path specified via option `--outdir` (default:
 `data/quac/results/test_project/analysis`).  Refer to the [system testing's
-output](./system_testing.md#expected-output-files) to learn more about the output directory structure. 
+output](./system_testing.md#expected-output-files) to learn more about the output directory structure.
 
 !!! tip
 
@@ -80,4 +61,3 @@ output](./system_testing.md#expected-output-files) to learn more about the outpu
 
     QuaC's output directory structure was designed based on the output structure of the [CGDS small variant caller
     pipeline](https://gitlab.rc.uab.edu/center-for-computational-genomics-and-data-science/sciops/pipelines/small_variant_caller_pipeline).
-
