+++ docs/input_output.md
@@ -33,9 +33,9 @@ Refer to our system testing directory for example sample config files at `.test/
 
 <!-- markdown-link-check-disable -->
 
-Samples belonging to a project are provided as input via `--pedigree` to QuaC in [pedigree file
-format](https://gatk.broadinstitute.org/hc/en-us/articles/360035531972-PED-Pedigree-format). Only the samples that are
-supplied in pedigree file will be processed by QuaC and all of these samples must belong to the same project.
+QuaC requires a [pedigree
+file](https://gatk.broadinstitute.org/hc/en-us/articles/360035531972-PED-Pedigree-format) as input via `--pedigree`.
+Samples listed in this file must correspond to those in sample config file (`--sample_config`).
 
 <!-- markdown-link-check-enable -->
 
