### 2023-10-06 e4755eb
expands output info
