### 2023-10-06 69f20b2
updates install test cmd
