### 2023-10-07 cac1a97
removes outdated note
