### 2023-10-07 uab-cgds-worthey/quac#86 @ManavalanG
@Redmar-van-den-Berg We have now implemented the requested change. Users can now provide input info in a config file via `--sample_config`. Documentation is [available here](https://quac.readthedocs.io/en/latest/input_output/) (please be sure use the `latest` version in readthedocs at the bottom right).
