### 2023-10-07 uab-cgds-worthey/quac#85 @ManavalanG
@Redmar-van-den-Berg I have updated docs to include a [section on editing quac-watch configs](https://quac.readthedocs.io/en/latest/quac_watch/#editing-configs). Please be to sure use the latest version in readthedocs at the bottom right.
