### 2023-10-07 #5313 @ManavalanG
@Redmar-van-den-Berg Thanks for taking the time to review our submission and provide feedback. We have made the requested changes.

1. Users can now provide samples input info in a config file via `--sample_config`. Documentation is [available here](https://quac.readthedocs.io/en/latest/input_output/). (https://github.com/uab-cgds-worthey/quac/issues/86)
2. We have updated docs to include a [section on editing quac-watch configs](https://quac.readthedocs.io/en/latest/quac_watch/#editing-configs). (https://github.com/uab-cgds-worthey/quac/issues/85)

Note: Please be sure to use the latest version in readthedocs at the bottom right.

I hope this resolves the issues. Please let me know if you have any questions :)
