### 2023-10-13 #5313 @ManavalanG
* I created a new release - [1.6](https://github.com/uab-cgds-worthey/quac/releases/tag/1.6)

> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
>
* [ ]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that e used in the JOSS paper.
