### 2023-10-16 #5313 @ManavalanG
@lpantano Zenodo is going through [a major upgrade](https://blog.zenodo.org/2023/10/13/2023-10-13-zenodo-rdm/), and it is not working fully as expected atm. QuaC's repo is already [part of zenodo](https://zenodo.org/records/10001881), but it doesn't let me update it to the recent QuaC version [1.6](https://github.com/uab-cgds-worthey/quac/releases/tag/1.6). As soon as zenodo is fully working again, I will update the version and will follow up here. Sorry about the delay!
